+++ src/Actions/ScanDirectories.php
@@ -4,25 +4,11 @@ declare(strict_types=1);
 
 namespace SebastiaanLuca\Module\Actions;
 
-use Illuminate\Filesystem\Filesystem;
 use Illuminate\Support\Collection;
 use SebastiaanLuca\Module\Entities\ModulesDirectory;
 
 class ScanDirectories
 {
-    /**
-     * @var \Illuminate\Contracts\Filesystem\Filesystem
-     */
-    private $files;
-
-    /**
-     * @param \Illuminate\Filesystem\Filesystem $files
-     */
-    public function __construct(Filesystem $files)
-    {
-        $this->files = $files;
-    }
-
     /**
      * @param \Illuminate\Support\Collection $directories
      *