+++ src/Commands/RefreshModules.php
@@ -35,11 +35,8 @@ class RefreshModules extends Command
             '--keep' => $this->option('keep'),
         ]);
 
-        $process = new Process(sprintf(
-            'cd %s && composer dumpautoload',
-            base_path()
-        ));
-
+        $process = new Process(['composer', 'dumpautoload']);
+        $process->setWorkingDirectory(base_path());
         $process->start();
 
         foreach ($process as $type => $data) {